+++ src/codec.rs
@@ -1,5 +1,5 @@
 //! Logic for all supported compression codecs in Avro.
-use crate::{errors::AvroResult, types::Value};
+use crate::{types::Value, AvroResult, Error};
 use libflate::deflate::{Decoder, Encoder};
 use std::io::{Read, Write};
 use strum_macros::{EnumString, IntoStaticStr};
@@ -34,17 +34,21 @@ impl Codec {
             Codec::Null => (),
             Codec::Deflate => {
                 let mut encoder = Encoder::new(Vec::new());
-                encoder.write_all(stream)?;
+                encoder.write_all(stream).map_err(Error::DeflateCompress)?;
                 // Deflate errors seem to just be io::Error
-                *stream = encoder.finish().into_result()?;
+                *stream = encoder
+                    .finish()
+                    .into_result()
+                    .map_err(Error::DeflateCompressFinish)?;
             }
             #[cfg(feature = "snappy")]
             Codec::Snappy => {
                 use byteorder::ByteOrder;
 
                 let mut encoded: Vec<u8> = vec![0; snap::max_compress_len(stream.len())];
-                let compressed_size =
-                    snap::Encoder::new().compress(&stream[..], &mut encoded[..])?;
+                let compressed_size = snap::Encoder::new()
+                    .compress(&stream[..], &mut encoded[..])
+                    .map_err(Error::SnappyCompress)?;
 
                 let crc = crc::crc32::checksum_ieee(&stream[..]);
                 byteorder::BigEndian::write_u32(&mut encoded[compressed_size..], crc);
@@ -64,26 +68,27 @@ impl Codec {
             Codec::Deflate => {
                 let mut decoded = Vec::new();
                 let mut decoder = Decoder::new(&stream[..]);
-                decoder.read_to_end(&mut decoded)?;
+                decoder
+                    .read_to_end(&mut decoded)
+                    .map_err(Error::DeflateDecompress)?;
                 decoded
             }
             #[cfg(feature = "snappy")]
             Codec::Snappy => {
-                use crate::Error;
                 use byteorder::ByteOrder;
 
-                let decompressed_size = snap::decompress_len(&stream[..stream.len() - 4])?;
+                let decompressed_size = snap::decompress_len(&stream[..stream.len() - 4])
+                    .map_err(Error::GetSnappyDecompressLen)?;
                 let mut decoded = vec![0; decompressed_size];
-                snap::Decoder::new().decompress(&stream[..stream.len() - 4], &mut decoded[..])?;
+                snap::Decoder::new()
+                    .decompress(&stream[..stream.len() - 4], &mut decoded[..])
+                    .map_err(Error::SnappyDecompress)?;
 
-                let expected_crc = byteorder::BigEndian::read_u32(&stream[stream.len() - 4..]);
-                let actual_crc = crc::crc32::checksum_ieee(&decoded);
+                let expected = byteorder::BigEndian::read_u32(&stream[stream.len() - 4..]);
+                let actual = crc::crc32::checksum_ieee(&decoded);
 
-                if expected_crc != actual_crc {
-                    return Err(Error::SnappyCrcError {
-                        expected: expected_crc,
-                        found: actual_crc,
-                    });
+                if expected != actual {
+                    return Err(Error::SnappyCrc32 { expected, actual });
                 }
                 decoded
             }